requirements.txt
streamlit>=1.25.0
langchain>=0.1.0
langchain-community>=0.0.35
python-docx
PyPDF2
